requirements.txt
azure-ai-documentintelligence>=1.0.0
openai>=1.0.0
pymupdf>=1.23.0
Pillow>=9.0.0
python-dotenv>=1.1.0
pdf2image>=1.16.0